requirements.txt
openai>=1.0.0
openai-agents>=0.6.0
qdrant-client>=1.8.0
cohere>=4.9.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
transformers>=4.35.0
numpy>=1.24.0
pytest>=7.4.0